Restore mysql index hint
We previously included this mysql index hint under some circumstances
for both build and buildset queries.  It initially appeared to no
longer be necessary after recent table changes, however that does not
seem to be the case, and in fact, it is necessary for some simple
build table queries.

It does no longer appear to be necessary for the buildset table, so
it is not added back here.

When applied to MySQL 5.7, these hints result in the build and buildset
queries returning the first N rows rather than the last N rows as
expected.

We should be able to resolve that issue as well, but it will require
further changes.  We're going to merge this as a short-term fix while
we work on a better solution.  The Zuul community has been notified
that master is currently a work in progress and to temporarily avoid
continuous deployment under mysql/mariadb.

Change-Id: Ia6d962957af3cac87d174145e69571a405ed505b