fix(web): stop toggling builds on clicks
In the web UI, when clicking on a change ID of a buildset, the
event bubbles and the buildset will be toggled open/closed which
is frustrating when you just want to view the change.

This patch stops the event from propagating to avoid other things
from firing but keep opening the link.

Change-Id: Ib2fd1051d21830d06fb5516a67656147efb39c1d