Add --disable-pipelines option
This facilitaties the creation of a Zuul system with a running config
(that will be kept up to date as long as it receives events) but does
not run any jobs or make any reports.  This can be used in conjunction
with zuul-web to serve REST API requests for introspection, or to create
a standby Zuul system with a warmed config cache, or to support other
debugging techniques.

This change adds an extra assertion to the wait-for-init test since it
would be too similar otherwise.  It also adds some documentation for
wait-for-init (so that both similar options are documented) and support
for setting both options by environment variables for ease of use
in k8s environments.

Change-Id: I3ee83b08c8280066cfa0744f2e30e41edd0f364c