Add inherit-files attr to evaluate parent file matchers
When building more complex zuul job constructs,
it is impossible to utilize the `files` attribute
of a parent job to reduce the amount of regex
rules specified in child jobs. Without inheritence,
child jobs have to repeat rules for any common file
paths. This PR adds a new `inherit-files` attribute
which will make the evaluation take `files` and
`irrelevant-files` of the parent explicitly into
account.

Task: #49809
Change-Id: I847944b77fb67f9ea1cb4e688994e08de9f59a98
Signed-off-by: Gondermann <gondermann@b1-systems.de>