Refresh git-dependent changes and add force_update_parent_dependency setting
When a gerrit change is merged, the dependencies are not always
updated correctly until you change or rebase a patch. Refreshing
the change gets around this issue to make sure the dependency
is merged. Consider the following scenario (merge-mode: cherry-pick):

A user creates relation chain patch A and B. They update only A
with a new patchset and gate. Then they apply the correct votes
for B to gate. If the change isn't refreshed, change B will not start
gating because it's dependencies are using an older patchset to
check is_merged.

Merge-mode cherry-pick supports merging a change that is an indirect
ancestor (parent is any patchset of another change) as long as there
are no conflicts with what is already merged in the repository.
Therefore, a new setting force_update_parent_dependency is added
to force the parent of a change to be the latest patchset to
corretly handle this case.

Change-Id: I6afc8b9897a0c04b416edd26b880386963932082