Fix callback detection in freeze playbooks
Ansible returns exit code 4 for both parse errors and unreachable hosts.
To disambiguate the two conditions, we use a callback plugin that is
run in the case of unreachable hosts.  However, we currently disable
all callback plugins during the freeze playbook in order to avoid
streaming output from that playbook.  That means that any unreachable
hosts in our very first playbook won't be detected as unreachable
(instead, Zuul will erroneously report a parse error.

(Note, that most unreachable hosts would hit this error during the setup
playbook, which does include all callback plugins, however, that's only
run for ssh hosts, so kubectl or network device hosts won't hit an
unreachable error until the freeze playbook.  We don't run the setup
phase for that because there are some use cases (ie, network devices)
where users expect the hosts to be unreachable early in the job.  And we
ignore the return code from the freeze playbook for the same reason. Only
once the job actually starts to run "real" playbooks that attempt to connect
to hosts do we actually act on unreachability.)

This change enables the unreachable host detection callback during all
playbooks (but continues to disable the output callbacks during freeze) so
that the errors reported to users in the job log reflect what's actually
happening.  Because we ignore the freeze playbook return code, this change
does not otherwise alter job results.

Change-Id: Ifa094765b1363b68af607f2cc6d75a3e5bf3e21d