Add ignore-setup-failure job option
We ignore unreachable hosts during the setup phase in order to support
a use case where a job is given a network appliance from nodepool, and
it is initially inaccessible, but the job itself enables access to it.

However, if a kubernetes pod is unreachable at the start of a job, that
should be an error.

Since it's more likely that we expect resources to be accessible at
the start of the job, let's change the default behavior to expect that,
but add an option to explictly enable the old behavior.

Change-Id: I6a3d6811b697cdce575c5a4b22bec9a087972c7b