Add zuul.build_refs variable
This variable is mostly intended for retroactive debugging of jobs.
There is currently no way to determine for which specific refs a
deduplicated job is being run.  This records that information in
the inventory file for the job.

This change also corrects the docs to indicate that change_url is
always present.

Change-Id: I990ecfe2a4b455fa800906ed0e542a3581bb8b29