Ignore /COMMIT_MSG in files matchers even more
We have mostly managed to ignore the /COMMIT_MSG in files matchers
(because it is unintuitive that it would be considered), but the recent
change to allow negated regexes in irrelevant-files exposed the fact
that it can still have an effect in that case.  The new feature hasn't
been used yet, so we could silently correct this, but a very similar
construction would be possible with the deprecated style of regex with
a negative lookahead.  Just in case someone wrote that, this change
includes a release note letting them know they can drop the /COMMIT_MSG
from their regex.

Change-Id: Ide04ed01224b5c0c48ab8d3c15ea7aef324cc42d