Visualize branches in ChangeQueues
The old status page visualized failing queue items (e.g. merge
conflicts) by creating a new "branch" in the change queue.

This change implements a similar functionality in the
PipelineDetailsView by nesting ProgressSteppers in a tree-like view to
visualize multiple branches of QueueItems within a ChangeQueue.

By utilizing a ProgressStepper for this purpose we can get rid of most
of the custom CSS code that was necessary to draw and connect the
"branching lines" in the old approach. We only need a little bit of CSS
to connect an inner ProgressStepper to the outer one.

In case a queue has multiple heads, each head will be rendered in a
separate tree (ProgressStepper).

Change-Id: Iee80b8d472cfbd8436301dff83edde5924bb7d08