Gitlab: make change.files persistent
Gitlab does not include files lists with merge request events, so
we always have Zuul perform that calculation itself.  But if a
merge request is updated after the fileschanges job completes, we
would overwrite our files list with None again.

Instead, recognize that for a given change+patchset, files are
immutable and don't update it if it's set already.

This matches the logic in Github, which, although it does provide
files with PR events, it only returns the first 300, so this logic
is needed in those cases.

Change-Id: I115c69b97e17cd4b01e3fa6c70140add6254283d