Implement admin actions (promote, dequeue) in new QueueItem component
So far, the new QueueItem component used in the PipelineDetailsView was
missing the implementation for the admin actions (promote and dequeue).

This change implements those admin actions. The implemenation itself is
taken from the old "Item" (containers.status.Item) component with a few
adaptions to the new component structure.

The admin actions are now only shown if the user is properly
authenticated.

Change-Id: I0768d04accca4e6580c8755dab8c4a8e600aadb7