Use proper background for Fetching and EmptyPage components
Currently, the "Fetching info..." string is always shown on a gray
background, while the rest of the UI uses a white background (or
black(ish) if darkmode is selected). This looks especially odd, when
the page changes from fetching to non-fetching state.

To fix this, wrap the Fetching component into a PageSection, so that
Patternfly choses the correct background color based on the selected
theme.

The same applies to the EmptyPage component.

I've noticed that there is no need to set a PageSectionVariants.dark in
case darkMode is enabled, as Patternfly will automatically switch the
background and font color accordingly for PageSectionVariants.light if
dark mode is enabled.

Change-Id: I9f47f1f0672ca574125d87ad81f671227db19322