Show zuul version in footer section on status page
This adds a dark footer section (based on the design of the Patternfly
homepage [1]) to the Zuul status page to show the deployed Zuul version.

Since the Zuul version is only available on the status page, the footer
is only shown on the status page. In general it might be a visual
improvement if we could show such a footer globally (on all pages).

[1]: https://www.patternfly.org/

Change-Id: I2921310b52430e4f341a93fde62f6b1d4abe7fbc