Fix loop when api-root authentication is configured
When api-root authentication is configured, access of zuul/api/tenants
would fail with 401 before authentication is done. Before this fix,
however, the browser would continuesly sending the request before
it has a chance to redirect to the authenticate page, which results
in a endless loop in some browser (as tested in Firefox), or loops for
randomly a few seconds and finally redirected to the authentication
page (as tested in Chrome).

It is fixed that it calls fetchTenantsIfNeeded() only when authentication
is done or not needed (api-root authencation is not configured).

Change-Id: I2cc67c791d694f329cd48c09d81cdda452eff12c