Add python-is-python3 for arm builds
The current version of re-ansi is using and old outdated version of
bs-platform (which is renamed to rescript), that only bundles x86
versions of ninja. Therefore the package tries to build ninja from
source, however this fails since it relies on having a python3 in PATH
with the name 'python'.

Newer versions of rescript bundles arm versions of ninja as well.

See:
https://github.com/rescript-lang/rescript-compiler/issues/5179

Change-Id: Ibc29c3c5b983c62268730936ae173d6a143c1f76