Update patch set 3
Patch Set 3:

Can you put the commit message in a comment in the dockerfile so we know why we're doing that?

Also, the comment could probably explain more what this is, and why we need this for arm, and whether it's temporary or permanent, and if temporary, under what conditions it could be removed.

Patch-set: 3
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Attention: {"person_ident":"Gerrit User 29671 \u003c29671@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"ADD","reason":"Someone else replied on the change"}