Allow early configuration of fake gerrit in tests
A followup change is going to require that we set the submitWholeTopic
value in our fake gerrit at the start of the test setup sequence (instead
of after zuul has already started as we do now).  The best way to do that
is to add a test decorator so that we can set the value for the test methods
that require it.

Because we are plumbing an increasing number of what are effectively per-test
variables to many different places in the tests, let's centralize these values
so that we don't have to pass them all around explicitly.

This adds a TestConfig class which holds some test-specific variables, including
the global fake change database, and passes this new object around instead of
several individual values.

There is room for adding more variables in the future and simplifying the class
and function signatures of many methods, but this is a start.

Change-Id: I514744b1ecca536fd844ed28d29fac3b6ca73a04