Expand the query cache scope to encompass multiple events
The query cache added to the scheduler was previously scoped to
a single event (but across all pipelines for that event).  Expand
the scope so that it encompasses processing all of the pending
tenant trigger events.  This reduces the number of queries in
the case of a fast cycle storm:

Old:  queries = 5*count
New:  queries = 4*count + 1

100 changes gives 401 queries.

The case where a user uploads or updates changes one at a time is
unchanged.

Change-Id: I9c1c086a05c13a7a7c10d3318ea334fd7e3c8cd5