Make the query cache continuous
This moves the query cache into the scheduler and persists it
permanently (ie, across pipeline runs) except that each time we
encounter an event newer than the last time we cleared the cache,
we will clear it again.

This means we should start with an empty cache each time a new
event arrives, but if a bunch of events arrive in succession, we
will use the same cached values.

This improves performance in topic dependencies in the slow case
(ie, developers adding changes one at a time).  The number of queries
in this case will be:

  queries = 5*count
  100 changes gives 500 queries

This does not alter the performance of the fast case (when all
changes in a topic appear at once).  That remains:

  queries = 4*count +1
  100 changes gives 401 queries

Change-Id: Iede8dc3d716279a07310726c647d8646e52ecc98