Update patch set 1
Patch Set 1:

That's a good idea, but I've been thinking about this and I know there are certain uses where having a standalone script would be more convenient than zuul-client.

For example, booting a new cloud node to replicate a failed job -- it would be easier to just curl the script and run it on the node.

If a developer already has zuul-client set up locally (or has docker set up and is familiar with running zuul-client in docker) then it's probably marginally more convenient to have it in zuul-client.  At least with docker, it's a one-liner (it's a long line, but still, one line to copy-paste).

But if they don't have that environment (or are not familiar with it) then it's probably easier to curl and run.

So overall, maybe there's a slight edge for the curl case?

We could do both: we could move this to zuul-client and keep it able to be run as a standalone file, then import it and wrap it with a zuul-client command.  Thoughts on that?

Patch-set: 1
Attention: {"person_ident":"Gerrit User 1 \u003c1@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"REMOVE","reason":"removed on reply"}