Abort pipeline processing for pending reconfigs earlier
Schedulers try to exit pipeline processing early when another
scheduler is trying to aquire the tenant write lock in order to
perform a tenant reconfiguration.  That's because that other
scheduler must sit idle until it can aquire the lock and proceed.
If there are many events or deep pipelines to process, the schedulers
that do hold the tenant lock may take quite some time to finish,
thus delaying the reconfiguration and the work of the scheduler
waiting to perform it.

We currently check for pending reconfigurations only once: between
locking each pipeline for processing.

We can check more often: before each pipeline event that we process,
and even between each queue item that we process.  In both of those
cases, it's okay to leave work partially completed; the next pipeline
run will finish the work.

However, checking for a pending reconfiguration involves some ZK,
traffic -- we must list each of the lock contenders, then fetch its
znode to see if it is for a pending reconfig.

In order to perform this check more often without significant
additional ZK traffic, we will set a child watch on the lock and read
the znode of any contendors and note their names.  Then when we are
ready to check for pending reconfigs, we will just check our local
data to see if we have seen a pending reconfig lock contender.

This will remove the ZK traffic from the current explicit checks before
each pipeline process, but will add traffic from the watcher if there
are many attempts to lock a pipeline in progress.

Change-Id: I8849905fd630d6362ff7dd2dbd6e76ca45b63c55