Store BuildSet.dependent_changes as change refs
When we set the configuration of a buildset, we collect all of
the dependent changes and serialize them so the relevant change
inforamtion is encoded in the buildset object.

We do this because changes at the head of the pipeline might
merge while we are processing and we still want to include them.

The change.toDict method includes the change's commit message.
In the case of a very large commit message, or a large number of
dependent changes with moderately sized commit messages, we can
exceed the 1MB zk limit.

To address this, store the dpendent changes list as a list of
cache key refs and then resolve those changes at the time that we
launch a job.

Change-Id: I3de7f11d6cd5d66eb383f4b962613a1d54365660