Fix race in test_timer_multi_scheduler
This test was failing to settle on a busy/slow system.  We don't
actually need to run the timer-triggered job in this test, we only
check the job registration.  So configure it to run less often so
that it is more likely that the test framework will be able to
settle.

(The only thing changed in the new test fixture is:

        - time: '* * * * * */1 1'
to
        - time: '* * * * * 1 1'
)

Change-Id: I8a69f3bfabf8faab32597e5304c3f681b15c95cd