Use latest patchset with Gerrit cherry-pick
Gerrit normally will not allow a change to merge if one if its merged
dependencies is not the latest patchset.  However, if the change is to
be cherry-picked, then Gerrit will allow this.  This is important
because when a change is cherry-picked, it always gets a new patchset
at the time of merging.  Gerrit does not, however, update any change's
dependencies.  So if a change points to an old patchset of a
dependency, then Zuul needs to know to consult the latest patchset to
determine whether a dependency has merged.

There are two cases to consider.  The first is a simple series of
changes where the first change merges (and in doing so, a new patchset
is created) and then the second is enqueued into gate.  One might
expect the second to fail to be enqueued because of the behavior
described above, however, when Zuul merges the change, it does not
notice the patchset update, so it marks the old patchset as merged
internally, and that allows the second change to proceed.  This is a
useful intentional bug (as it saves a lot of extra work), so we keep
it.

The second case is when a user adds a patchset to the first change
before it merges.  This forces the issue because the second change
depends on patchset 1, but patchset 2 was enqueued and patchset 3
merged.  That means the intentional bug described above doesn't help
us since it is erroneously marking patchset 2 as merged, but the
second change depends on patchset 1.  To correct this, we need to do
more work.

This change adjusts Zuul's internal dependency pointers for gerrit
git-dependent changes to use the latest patchset iff a change is being
cherry-picked.  In the case above, this would cause the second change
to depend on the third patchset of the first change, and Zuul will
query the change anew in that case (since it has never seen the third
patchset) and discover that it is merged.

Notably, the ssh code path does not provide enough information for us
to do this (only http gerrit queries do), so the intentional bug above
is important to keep in place so that we don't break the more typical
case in the ssh code path.

Change-Id: Icc6ee727654139c589df3768254178a4bc6e326f