Fix missing unsafe marker in new build_refs var
The recent change to add build_refs added a new place where we include
the commit message in a zuul variable, but did not mark it unsafe,
which means ansible may try to template it.  Commit messages frequently
have character sequences in them that look like jinja templates, so
for convenience, we mark them !unsafe to stop ansible from templating
them.

Correct the missing !unsafe marker, and add a test that verifies the
commit message does not appear in the vars unmarked.

Change-Id: I45ebed779884cab32718064f40716581854ea03a