Run py312 jobs on Ubuntu Noble
Ubuntu Noble packages python 3.12 for us. Rather than use pyenv on
Ubuntu Jammy to build python 3.12 which adds time to jobs (and probably
gives us a less efficient python build) we switch the job to run on
Ubuntu Noble using the packaged python instead.

We update bindep.txt to force the installation of libffi8 on Noble
otherwise it falls back to libffi6 which is not available.

Change-Id: I274d8e922f448be5bc7e3127e66b4bd246eb02d9