Add filtering support for pipeline status views
This adds a filtering mechanism for the pipeline overview (status) and
pipeline details pages. It reuses the filter toolbar used on other pages
(Builds, Buildsets) to keep the user interface consistent. However, in
this case, the filtering is done on the client side. Also, wildcard
searches (with the asterisk char) are supported as they are on the
Builds and Buildsets pages.

Displaying empty pipelines via the "show all pipelines" toggle might
clash with a filtered view. We simply disable the toggle if a filter is
set, so these view settings are mutually exclusive.

Change-Id: Ibd6a0cef837b876fe3e20d42149a55b2134d8eda