Pre-filter gerrit events based on triggers
There is a certain amount of overhead involved in sending a Gerrit
event to be processed by a scheduler.  In recognition of this,
we already discard some Gerrit events as soon as we receive them
(events of certain types that are not relevant to Zuul, or events
related to notedb operations).

In addition to these Gerrit-related events, it's also possible for
users to have their own ref spaces in Gerrit that are used for
purposes unrelated to Zuul.  In order to ignore these events, this
change adds pre-filtering to Gerrit connections.

Now when we receive a Gerrit event, we will briefly compare it to
all of the configured pipeline triggers.  If it matches some simple
criteria (event type, branch, ref pattern), then we will keep the
event and process it as normal (which will entail sending it to
the scheduler event queue for distribution to pipelines and the
normal method of fully matching the event to a pipeline).  If it
doesn't match this high-level criteria, we will ignore it and avoid
the overhead of submitting it to ZK and the scheduler queue.

This may use a small amount of additional scheduler CPU in the initial
gerrit event loop, but can save significant IO and network traffic
by avoiding fully populating the event structure and submitting it
to ZK.

Change-Id: I4f844f7c5abf0d453c7825357a22abe4cdc566d8